rename baseline_model to consensus_pledge_model
