+++ .gitignore
@@ -130,3 +130,6 @@ dmypy.json
 
 # .DS_Store
 .DS_Store
+
+
+data/simulations/*.pkl.gz
