+++ README.md
@@ -3,12 +3,12 @@ Interactive Calculator for the economic incentives around the Filecoin Baseline 
 
 ## How to run it
 
-- Option 1 (CLI): Just pass `python -m baseline_model`
+- Option 1 (CLI): Just pass `python -m consensus_pledge_model`
 This will generate an pickled file at `data/simulations/` using the default single run
 system parameters & initial state.
-    - To perform a multiple run, pass `python -m baseline_model -e`
-- Option 2 (cadCAD-tools easy run method): Import the objects at `baseline_model/__init__.py`
-and use them as arguments to the `cadCAD_tools.execution.easy_run` method. Refer to `baseline_model/__main__.py` to an example.
+    - To perform a multiple run, pass `python -m consensus_pledge_model -e`
+- Option 2 (cadCAD-tools easy run method): Import the objects at `consensus_pledge_model/__init__.py`
+and use them as arguments to the `cadCAD_tools.execution.easy_run` method. Refer to `consensus_pledge_model/__main__.py` to an example.
 - Option 3 (Streamlit, local)
 - Option 4 (Streamlit, cloud)
     1. Fork the repo
@@ -19,7 +19,7 @@ and use them as arguments to the `cadCAD_tools.execution.easy_run` method. Refer
 ## File structure
 
 - `app/`: The `streamlit` app
-- `baseline_model/`: the `cadCAD` model as encapsulated by a Python Module
+- `consensus_pledge_model/`: the `cadCAD` model as encapsulated by a Python Module
 - `data/`: Simulation / Post-processed datasets
 - `notebooks/`: 
 - `scripts/`:
