+++ consensus_pledge_model/__init__.py
@@ -1,5 +1,5 @@
-from baseline_model.params import PARAMS, INITIAL_STATE, TIMESTEPS, SAMPLES
-from baseline_model.structure import BLOCKS
+from consensus_pledge_model.params import PARAMS, INITIAL_STATE, TIMESTEPS, SAMPLES
+from consensus_pledge_model.structure import BLOCKS
 
 default_run_args = (INITIAL_STATE,
                     PARAMS,
